DORN-53: Fixed issue where user is logged out but not redirected to login page.
